refactor: hook file structure
