+++ src/hooks/service/trending/index.ts
